pre_qualification/research_engine.py
import logging
import requests
from bs4 import BeautifulSoup
from typing import List, Dict

logger = logging.getLogger(__name__)

class ResearchEngine:
    def __init__(self):
        self.supported_research_types = ['company_financials', 'decision_maker_background', 'recent_news', 'tech_stack_analysis']
        logger.info("Initialized Pre-Qualification Research Engine")

    def _fetch_company_financials(self, company_domain: str) -> Dict:
        """Placeholder for gathering company financial indicators (revenue, funding, growth)"""
        # Will integrate with financial APIs (e.g., Crunchbase, PitchBook) in later iterations
        logger.info(f"Fetching financials for {company_domain}")
        return {"revenue_range": "$10M-$50M", "funding_stage": "Series B", "year_over_year_growth": 25}

    def _fetch_decision_maker_background(self, contact_name: str, company_domain: str) -> Dict:
        """Placeholder for gathering decision-maker professional history (LinkedIn, company bios)"""
        # Will integrate with LinkedIn API or web scraping in later iterations
        logger.info(f"Fetching background for {contact_name} at {company_domain}")
        return {"title": "VP of Sales", "tenure": "3 years", "previous_company": "Tech Growth Inc.", "decision_authority": "High"}

    def _fetch_recent_news(self, company_domain: str) -> List[Dict]:
        """Placeholder for gathering recent company news (product launches, partnerships, press releases)"""
        # Will integrate with news APIs (e.g., NewsAPI) or web scraping in later iterations
        logger.info(f"Fetching recent news for {company_domain}")
        return [
            {"headline": "Company launches new enterprise platform", "date": "2025-10-15", "source": "TechCrunch"},
            {"headline": "Company secures strategic partnership", "date": "2025-09-20", "source": "Business Insider"}
        ]

    def _analyze_tech_stack(self, company_domain: str) -> List[str]:
        """Placeholder for analyzing company tech stack (websites, job postings)"""
        # Will use web scraping with BeautifulSoup or tools like BuiltWith in later iterations
        logger.info(f"Analyzing tech stack for {company_domain}")
        return ["Python", "React", "AWS", "Salesforce"]

    def run(self, leads: List[Dict]) -> List[Dict]:
        """Execute pre-qualification research on processed leads"""
        logger.info(f"Starting pre-qualification research for {len(leads)} leads")
        researched_leads = []
        for lead in leads:
            if not lead.get("domain"):
                logger.warning(f"Skipping research for {lead.get('company_name')} - no domain found")
                researched_leads.append(lead)
                continue

            # Gather research data
            lead["financials"] = self._fetch_company_financials(lead["domain"])
            lead["decision_maker_background"] = self._fetch_decision_maker_background(lead.get("contact_name", "Unknown"), lead["domain"])
            lead["recent_news"] = self._fetch_recent_news(lead["domain"])
            lead["tech_stack_analysis"] = self._analyze_tech_stack(lead["domain"])

            researched_leads.append(lead)

        logger.info(f"Pre-qualification research complete for {len(researched_leads)} leads")
        return researched_leads