requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
elevenlabs>=0.30.0
ruff>=0.6.0
simple_salesforce>=1.0.0
serpapi>=0.1.0